requirements.txt
requests
pandas
faker
uuid
psycopg2-binary
tqdm
python-dotenv
fastapi
uvicorn
matplotlib
sqlalchemy
aiosmtplib
twilio
apscheduler
python-jose
passlib[bcrypt]
psycopg2-binary
argon2-cffi